+++ RELEASE_NOTES.md
@@ -1,4 +1,12 @@
-# app-framework release notes
+# create-strato-dapp release notes
+
+### Version 0.4.0
+
+- Renamed app-framework to create-strato-dapp
+- Converted the tool into npm package
+- Added the programmatic service token fetch during deploy which makes the ADMIN_TOKEN env var obsolete
+- Removed the blockapps-sol git submodule dependency from generated app
+- Added basic tests in generated app
 
 ### Version 0.3.0